redirect: move View Implementation
