+++ nextstep-mvc/src/main/java/nextstep/mvc/tobe/view/JsonView.java
@@ -1,17 +1,17 @@
-package nextstep.mvc.tobe;
+package nextstep.mvc.tobe.view;
+
+import com.fasterxml.jackson.databind.ObjectMapper;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
+import java.io.PrintWriter;
 import java.util.Map;
 
 public class JsonView implements View {
     @Override
     public void render(Map<String, ?> model, HttpServletRequest request, HttpServletResponse response) throws Exception {
-
-    }
-
-    @Override
-    public String getName() {
-        return null;
+        ObjectMapper objectMapper = new ObjectMapper();
+        PrintWriter out = response.getWriter();
+        out.write(objectMapper.writeValueAsString(model));
     }
 }
