+++ nextstep-mvc/src/main/java/nextstep/mvc/tobe/view/View.java
@@ -1,4 +1,4 @@
-package nextstep.mvc.tobe;
+package nextstep.mvc.tobe.view;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
@@ -6,6 +6,4 @@ import java.util.Map;
 
 public interface View {
     void render(Map<String, ?> model, HttpServletRequest request, HttpServletResponse response) throws Exception;
-
-    String getName();
 }
